refactor: Period components name consistency
